requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
orjson==3.10.7
passlib==1.7.4
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.10
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
typer==0.12.5
types-passlib==1.7.7.20240819
types-psycopg2==2.9.21.20241019
typing==3.7.4.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
